Move code for track-to-collision association in dedicated class
