+++ Common/DataModel/CMakeLists.txt
@@ -16,4 +16,5 @@ o2physics_add_header_only_library(DataModel
                                           FT0Corrected.h
                                           Multiplicity.h
                                           PIDResponse.h
+                                          CollisionAssociationTables.h
                                           TrackSelectionTables.h)
