+++ PWGHF/TableProducer/candidateCreatorB0.cxx
@@ -19,7 +19,7 @@
 #include "Framework/AnalysisTask.h"
 #include "DCAFitter/DCAFitterN.h"
 #include "Common/Core/trackUtilities.h"
-#include "Common/DataModel/CollisionAssociation.h"
+#include "Common/DataModel/CollisionAssociationTables.h"
 #include "ReconstructionDataFormats/DCA.h"
 #include "ReconstructionDataFormats/V0.h"
 #include "PWGHF/DataModel/CandidateReconstructionTables.h"
