+++ PWGHF/TableProducer/candidateCreatorBplus.cxx
@@ -19,7 +19,7 @@
 /// \author Antonio Palasciano <antonio.palasciano@cern.ch>, Università degli Studi di Bari & INFN, Sezione di Bari
 
 #include "Common/Core/trackUtilities.h"
-#include "Common/DataModel/CollisionAssociation.h"
+#include "Common/DataModel/CollisionAssociationTables.h"
 #include "DCAFitter/DCAFitterN.h"
 #include "Framework/runDataProcessing.h"
 #include "Framework/AnalysisTask.h"
